requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
snowflake-connector-python==3.14.0
gunicorn==21.2.0
whitenoise==6.6.0

